premium set of golf clubs (4 times) a Broadway show ticket (mid-range) (4 times) a basic tablet computer (3 times) a designer handbag (3 times) a smart watch (3 times)
thats an example for $23k and it works but id rather like 10 or 12 different things multiple times i think it would look cooler so just add more, a tonnn more

Also ensure the perspective calculations are calculated before the page fully pops up, theres a bug where it switches around. It should be an array with all the things that you add. So randomly pull from the list, subtract from the time from it. If its negative then pick something else til its not. Get as close to the things you pick equalling the time 98% minimum or so. Do this for both