Woah wtf this looks terrible. First off it should be something like (keyword like, you have creative freedom) (semi colon means different tab)
So the way a lot of apps have been doing stuff is with those little bubble filters. They kind of look like little bubbles. You can click on them, and you can add them to your selection, and they're really great little UI tools. So when you go in the Explorer for their actions, by default, one of the things that's bubbled is, like, your matches or something along those lines that will basically just be the things that the only filter is they have at least one trait in common with what you have put down. You can select that off by clicking the little X to the right, because when you click on something a little, it fills in. So it goes from, like, light to dark, or in dark mode from, like, dark to, like, a lighter color, or something. I don't know, make it work. And you're going to go ahead and you have that, so you can unselect that. If you unselect that, it gives you a list of literally all of the organizations that have any reference to you. And also, make sure you add more. There should be around about 50 or so. Just get as many as you can. And there should be ones for, like, everything that you can think of that you can, like, click on. And just, like, a little bit about, like, what it's trying to do, so it's nice and easy. Uh, it should be better. Also, the scroll bar on the Advocates For You review of, like, how it aligns with your concerns, that little tooltip, there is no scroll bar. So sometimes it will, like, spill out of the box, and you can't scroll down to read the full thing, which obviously is a bit of a problem. So please try to fix that. Um, if only, only add a scroll bar if a scroll bar is needed, obviously. Um, or just if the scroll bar isn't a good UI element, you think, then go for something else. So yeah, definitely fix the UI there and make it better overall. Add more organizations. Try to add even more badges. The Broadfolks, the Niche, those are very great. They're really smart, and I love those. Um, try to do just more, um, your end goal is to encourage it, to make each thing sound like it has its own merits, which it does in real life. So that's kind of your goal there.